backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
pymupdf==1.24.3
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
faiss-cpu==1.8.0
sentence-transformers==3.0.0
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1
aiofiles==23.2.1
sse-starlette==2.1.0
tiktoken==0.7.0
